archive/various/pyproject.toml
[project]
name = "tabularshenanigans"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogluon>=1.4.1b20251203",
    "holidays>=0.78",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.8",
    "kaggle>=1.7.4.5",
    "matplotlib>=3.10.3",
    "mlflow>=3.3.1",
    "optuna>=4.5.0",
    "pandas>=2.3.0",
    "polars>=1.35.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.43",
    "torch>=2.7.1",
    "xgboost>=3.0.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
